+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Versions and bullets are arranged chronologically from latest to oldest.
 
+## Unreleased
+
+- Disabled the additional pre-authentication step by default. Set `$wgUnifiedExtensionForFemiwikiPreAuth` to true to enable it again.
+
 ## v2.0.0
 
 - Queries more fields of RelatedArticles to fix cache.