+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 Versions and bullets are arranged chronologically from latest to oldest.
 
-## Unreleased
+## v3.0.0
 
-- Disabled the additional pre-authentication step by default. Set `$wgUnifiedExtensionForFemiwikiPreAuth` to true to enable it again.
+- BREAKING CHANGE: Disabled the additional pre-authentication step by default. Set `$wgUnifiedExtensionForFemiwikiPreAuth` to true to enable it again.
 - Introduces `$wgUnifiedExtensionForFemiwikiSoftDefaultOptions` which is used to change the default options only for the new users or reset options.
 - Makes HelpPanel and Homepage are set when after reset if configured for new users.